Add a formats option for the cache task
I wish lein v would output the version file to an EDN data structure for easy
consumption in Leiningen tasks, also I wanted to display the application's
version by embedding the version in the code bundle in CLJS for instance.

So I add a formats variadic option to the cache function and task allowing a
variable list of formats to be supplied to the cache function. The choosen
suffix (edn or cljx for instance) will deduce the content of the file (EDN or
clojure source code).

See the documentation for the `lein v cache` task in the README.md for complete
usages, examples and use cases.