Make sure the cache directory exists if necessary.
My users were having trouble building [Beat Link
Trigger](https://github.com/Deep-Symmetry/beat-link-trigger) from
source, it was mysteriously crashing complaining about `version.edn`
not existing, and it took a while to figure out why. That file is in
my `.gitignore` because it is a build product, and I had no other
files in the `resource` tree, so git simply did not create those
directories when someone cloned the project. Then `lein-v` failed to
write the file because the necessary parent directories did not exist
when it ran.

This one-liner should solve this problem in the future by creating the
parent directories of the configured cache directory when necessary.