Fix multiple git executable in Windows problem
In a situation with multiple git executables (git.exe) on the search
path, the underlying "where.exe" returns multiple matches in multiple
lines. This commit will fix the problem arising with this situation
where the git command to be called is this multi-line list. Instead, the
first git executable is returned find-windows-git.